fix: fix cxxtest bug

+++ cxxtest/cxxtest/ValueTraits.cpp
@@ -127,7 +127,7 @@ char *bytesToString(const unsigned char *bytes, unsigned numBytes, unsigned maxB
 unsigned ValueTraits<const double>::requiredDigitsOnLeft(double t)
 {
     unsigned digits = 1;
-    for (t = (t < 0.0) ? -t : t; t > 1.0; t /= BASE)
+    for (t = (t < 0.0) ? -t : t; t > 1.0; t /= static_cast<double>(BASE))
     {
         ++ digits;
     }
@@ -162,7 +162,7 @@ void ValueTraits<const double>::normalNumber(double t)
     s = copyString(s, ".");
     for (unsigned i = 0; i < DIGITS_ON_RIGHT; ++ i)
     {
-        s = numberToString((unsigned)(t *= BASE) % BASE, s);
+        s = numberToString((unsigned)(t *= static_cast<double>(BASE)) % BASE, s);
     }
 }
